Implementation/Shopping-Cart-MS/requirements.txt
Flask
Flask-Migrate
connexion[swagger-ui]
sqlalchemy
pyjwt
marshmallow-sqlalchemy
flask-bcrypt
python-consul